rename logicalAnswer
