+++ .env.example
@@ -12,9 +12,9 @@ DISKS_DIR="/public/uploads"
 DB_HOST="localhost"
 DB_PORT="3306"
 DB_CONNECTION="mysql"
-DB_DATABASE="ticket"
+DB_DATABASE="assessment"
 DB_USERNAME="root"
-DB_PASSWORD=""
+DB_PASSWORD="password"
 DB_SYNCHRONIZE=""
 DB_LOGGING="false"
 DB_STORAGE=
