+++ src/modules/game_result/logicalGameResult.service.test.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { LogicalGameResultService } from './services/logicalGameResult.service';
+import { LogicalGameResultService } from './services/logicalAnswer.service';
 import { GameResultService } from './services/gameResult.service';
 import { GameService } from '@modules/game/game.service';
 import { Repository } from 'typeorm';
