+++ src/modules/game_result/services/logicalAnswer.service.ts
